initial splitting out of code
